chore(lib): remove deprecated constructs
