--- apidocs/enumerations/CrawlerTargetType.md
@@ -1,19 +1,0 @@
-[**@cdklabs/generative-ai-cdk-constructs**](../README.md) • **Docs**
-
-***
-
-[@cdklabs/generative-ai-cdk-constructs](../README.md) / CrawlerTargetType
-
-# Enumeration: CrawlerTargetType
-
-## Enumeration Members
-
-### RSS\_FEED
-
-> **RSS\_FEED**: `"rss_feed"`
-
-***
-
-### WEBSITE
-
-> **WEBSITE**: `"website"`
