+++ src/common/base-class/construct-name-enum.ts
@@ -19,8 +19,6 @@ export enum ConstructName {
   HUGGINGFACESAGEMAKERENDPOINT = 'C6',
   JUMPSTARTSAGEMAKERENDPOINT = 'C7',
   AWSCONTENTGENAPPSYNCLAMBDA = 'C8',
-  AWSRAGAPPSYNCSTEPFNKENDRA = 'C9',
-  AWSWEBCRAWLER = 'C10',
   AWSTEXTTOSQL= 'C11',
   AWSLLAMAINDEXDATALOADER = 'C12',
 }
