+++ src/index.ts
@@ -16,7 +16,6 @@ export * from './patterns/gen-ai/aws-langchain-common-layer';
 export * from './patterns/gen-ai/aws-qa-appsync-opensearch';
 export * from './patterns/gen-ai/aws-model-deployment-sagemaker';
 export * from './patterns/gen-ai/aws-contentgen-appsync-lambda';
-export * from './patterns/gen-ai/aws-web-crawler';
 export * from './patterns/gen-ai/aws-text-to-sql';
